+++ packages/dsi-logo-maker/functions/fillSVG/index.js
@@ -8,7 +8,10 @@ import {
 } from "../../utils/blocks";
 import { Block } from "../../utils/blocks-components";
 
-export const handler = ({ width, height, done, logoWidth, logoRatio }) => {
+export const handler = ({ params, mechanic }) => {
+  const { width, height, logoWidth, logoRatio } = params;
+  const { done } = mechanic;
+
   const rows = 2;
   const cols = 13;
   const logoHeight = Math.floor((logoWidth / logoRatio) * rows);