+++ packages/create-mechanic/function-blank/index.js
@@ -1,4 +1,4 @@
-export const handler = () => {};
+export const handler = ({ params, mechanic }) => {};
 
 export const params = {};