+++ packages/dsi-logo-maker/functions/textAndImageCanvas/index.js
@@ -6,7 +6,7 @@ import {
 } from "../../utils/blocks";
 import { drawBlock } from "../../utils/blocks-canvas";
 
-export const handler = (params, mechanic) => {
+export const handler = ({ params, mechanic }) => {
   const {
     width,
     ratio,